Fixes some cyclic runtime dependencies
